count_colonies is a simple python script for counting round bacterial colonies
It assumes a square image.

It assumes the colonies are on a petri dish which both is centered in the
image and takes up the entire image (the script makes a round mask within the square image).

It assumes the colonies are brighter than the agar.

It depends on skimage, numpy, and matplotlib

USAGE:  python count_colonies.py [imagename.tif]

OPTIONAL ARGUMENT: no_plot (turns off plotting)

OUTPUT:  A text file with the x/y locations of the found colonies